added startup script and killing only if codecombat runs

+++ bin/stop_server.sh
@@ -4,8 +4,14 @@ cd "`dirname \"$0\"`"
 
 if [ -f codecombat.pid ]
 then
-  sudo kill -9 `cat codecombat.pid`
-  echo "killed codecombat server with pid "`cat codecombat.pid`
+  if wget -O - http://locahost:3000 >>/dev/null
+  then
+    echo "codecombat running on port 3000"
+    sudo kill -9 `cat codecombat.pid`
+    echo "killed codecombat server with pid "`cat codecombat.pid`
+  else
+    echo "codecombat is not running"
+  fi
   rm codecombat.pid
 fi
